What if instead of setting a default, we just returned one?
Instead of always setting a default cookie in the root module when we're in a test environment, let's just return the default value in `getPrivacySettings()` so the privacy settings dialog can go about its business without us having to touch the cookies. That seems to actually make it a little easier to read and reason about, and we get rid of the module side-effects.