[WNMGDS-2658] Fix crash of privacy settings dialog in node environments (#2908)
* Avoid crash in non-browser environments

I was trying to test this in our unit tests, but it is difficult. I can change the jest environment on a per-module basis with https://jestjs.io/docs/configuration#testenvironment-string, but it still runs the `setupTests` module and crashes. It was taking too much time to try reconfiguring Jest, so I tried mocking like this:

```
let windowSpy;

function mockNodeEnvironment() {
  windowSpy = jest.spyOn(window, "window", "get");
  windowSpy.mockImplementation(() => undefined)
}

function restoreJsdomEnvironment() {
  windowSpy.mockRestore();
}
```

but even though I could get `window` to be `undefined`, I couldn't get the `typeof window === 'undefined'` to be true.

The reason this is a problem now and wasn't before is that previously the `PrivacySettingsDialog` wouldn't be rendered in build environments like Gatsby because that component's visibility was controlled by conditionally rendering it. When we switched to `isOpen`, it would render that dialog even if not open, which caused this code to be executed where it wasn't before.

* What if instead of setting a default, we just returned one?

Instead of always setting a default cookie in the root module when we're in a test environment, let's just return the default value in `getPrivacySettings()` so the privacy settings dialog can go about its business without us having to touch the cookies. That seems to actually make it a little easier to read and reason about, and we get rid of the module side-effects.