[WNMGDS-2669] Remove unique and redundant semantic variables for dialogs (#2933)
Remove unique and redundant semantic variables for dialogs

These aren't used by any other components, and arguably it's not a theme-level semantic variable. It's pretty specific to a component. It's conceivable that it could be used for multiple dialog-like components in the future, but it isn't right now, and it doesn't even differ across themes. It's just an unnecessary variable.