Fix `cdn-html-css`
Updating just the version fixed the inline error displaying inline (on the same line as the label); however, this markup is all out of date. Honestly, we don't really want people using plain HTML to make the forms when we could do it with web components in the future. The markup is too complex and prone to change.