[WNMGDS-2653] Examples to use local design system packages (#2938)
* Configure Lerna to manage the example projects too

so we can automatically use the latest version of the design system as a dependency

* Had to update the workspace config and deduplicate some package names

* Install from the root and update yarn lockfile

Now the examples are actually using the local version of our DS packages

* According to TypeScript, this returns a promise

* Maybe we don't need the package in lerna if it's in the workspace?

I don't want it to show up when we go to bump versions

* Maybe we can ignore them with this option

* Upgrade the node version we use

I'm trying to replace create-react-app-typescript because it's deprecated and is actually not compatible with linking our packages through workspaces, but what I'm trying to replace it with requires v18+. The version we were using (v16) is no longer maintained either.

* Start creating a new astro example to replace create-react-app

* Move the CRA TS examples into this one

* Disable the first-child rule for headings

* Web components page

* Don't need a label example. It's outdated anyway

* Update readmes

* Remove the `create-react-app-typescript` example

Replaced it with the `astro` example that is modern and doesn't use deprecated libraries

* Update webpack command and fixed issue caused by preact version mismatch

* Make all the (p)react-app examples uniform (and get them all to work)

* Update examples root README list

* Rein in the hacky override here

* Use native CSS nesting (new)

* Fix `cdn-web-components` adding machine

* Fix `cdn-html-css`

Updating just the version fixed the inline error displaying inline (on the same line as the label); however, this markup is all out of date. Honestly, we don't really want people using plain HTML to make the forms when we could do it with web components in the future. The markup is too complex and prone to change.

* Clean up the HTML/CSS example (see previous commit message)

* Replace deprecated css class

* Get rid of duplicate h1

* Simplify build instructions that I missed

* We don't need any local styles because we have a reset. Simplify!

* Discovered that our web component wrappers don't have the right display properties

* Update dialog example to use new API

* Make the Pagination example functional

* Add jsx config so we don't get a bunch of squiggly errors in our examples