This is the minimum version that causes the issue in Preact Storybook
To reproduce the error, go to the day picker story and click the calendar button. It will throw an error about `.current` and coming from a jsx function

I wonder if the change [between react-day-picker v8.8.2 and v8.9.0](https://github.com/gpbl/react-day-picker/compare/v8.8.2...v8.9.0) causing problems is the removal of the React imports from the modules. I know we've had issues with that in the past. Right now if I upgrade past v8.8.2, Storybook will work with React but not with Preact. I think in order for us to be able to use libraries that use the new JSX transform, we have to use the new JSX transform, but we can't use the new JSX transform until we can safely drop support for React 16.

I've tried changing the TS config setting to
```json
"jsx": "react-jsx",
```
but it doesn't seem to fix the problem.