Come up with a workaround to incorrect rehydration (client-only components)
And created a little helper function for it. Statically rendering them causes them to have to undergo rehydration which compares the statically rendered HTML to the HTML rendered by React in the client after the JavaScript loads. If there are differences that can't be reconciled, everything blows up. So can we skip that processes entirely for a node in the DOM/React tree? Yeah, let's just skip it for now and make it so they only render in a browser.