New private packages broke regex for tags, and I'm making it better
Along with fixing the regex pattern to make it able to not get tripped up by packages that don't have a `@cmsgov` namespace, I made it only match the four design system packages so we reduce the tag noise. This really only removes our private docs package from the tags, but I think that's an improvement.