Fix issue in `examples/react-app` to illustrate a point
We can use our updated library that uses the new JSX transform in a project that doesn't yet use the new JSX transform, but first I had to fix the example so it imported React. Previously it relied on the fact that our library defined `React` globally!

Now that the point has been illustrated, I can actually just update these examples to use the new JSX transform as well