[WNMGDS-2726] Update alert background colors to ensure sufficient color contrast (#3007)
* Update success and error alert backgrounds to new approved colors to fix contrast

These both had color-contrast issues with primary-colored links and such

* Update VRT snapshots

* These example VRTs changed too

* Don't need these variables that were used to correct contrast issues

This is the root issue the bg color change was trying to solve. There are too many possible combinations to account for with special CSS overrides. We instead chose to make sure the primary color worked on all normal background colors

* Put some links in the success and error alert stories to check the contrast