[WNMGDS-2757] Upgrade `react-day-picker` library (#3064)
* This is the minimum version that causes the issue in Preact Storybook

To reproduce the error, go to the day picker story and click the calendar button. It will throw an error about `.current` and coming from a jsx function

I wonder if the change [between react-day-picker v8.8.2 and v8.9.0](https://github.com/gpbl/react-day-picker/compare/v8.8.2...v8.9.0) causing problems is the removal of the React imports from the modules. I know we've had issues with that in the past. Right now if I upgrade past v8.8.2, Storybook will work with React but not with Preact. I think in order for us to be able to use libraries that use the new JSX transform, we have to use the new JSX transform, but we can't use the new JSX transform until we can safely drop support for React 16.

I've tried changing the TS config setting to
```json
"jsx": "react-jsx",
```
but it doesn't seem to fix the problem.

* Discovered that our SingleInputDateField doesn't work in native Preact projects

It only works if Preact is in React-compatability mode. It works in examples/preact-react-app but not examples/preact-app.

* Upgrade to the version of RDP that gives us date-fns 3.0 with ES Modules support

While this doesn't work in this commit, the whole goal of this effort is to make it work. The problem is figuring out how to get their built-in JSX transform functions to play nice with the way _we_ handle JSX, React, and Preact.

* Upgrade date-fns too. It all works in the Astro example

* Upgrade to fixed version

* Fix `preact-app` example

* Update CSS because of `aria-disabled` to `disabled` change upstream

* Fix selectors

* Address color-contrast issues with design changes

These changes were made after discussion with both Accessibility and Design and have been approved by UX leads

* Fix strike-through in forced-colors mode

* Update VRT refs

* Add it to the astro example too

* Remove from examples where it isn't needed now that we've tested it

* Update VRT refs of astro example where we added the date field

* Now that these buttons are actually disabled, we don't need to disable pointer events

And this keeps us from being able to apply the `not-allowed` cursor style