Add `ds-accordion-item` and make necessary changes to `AccordionItem`
The previous logic for determining whether it is controlled was incompatible with web components, because we always supply a callback for all events.

I also chose not to include `closeIconComponent` and `openIconComponent`, since those are really React-specific and we want to move away from that kind of customization anyway. Didn't include `isControlledOpen` either