[WNMGDS-2771] Add `ds-accordion` and `ds-accordion-item` web components (#3071)
* Add `ds-accordion-item` and make necessary changes to `AccordionItem`

The previous logic for determining whether it is controlled was incompatible with web components, because we always supply a callback for all events.

I also chose not to include `closeIconComponent` and `openIconComponent`, since those are really React-specific and we want to move away from that kind of customization anyway. Didn't include `isControlledOpen` either

* Add `ds-accordion`

* Add a story and fix styles

* Add story for items so we have the docs and args table

* Add snapshots for the storybook docs

* Add VRT snapshots

* Use each component folder's index file as the entry to fix accordion

The accordion bundle contains two components. At some point during development of the bundle-splitting logic, using the index file wasn't working, but now it seems to be working.

* Update example VRTs

* Add missing event documentation

* These doc snapshots changed