[WNMGDS-2746][WNMGDS-2776] Update token build scripts to use JSON source (#2952)
* This script can just be about exporting CSS/Sass now

After the Figma changes, we have a different way of storing tokens, and they're all in one folder. We haven't needed the multiple input directories for a while, and we don't really have a big reason to run the command multiple times for different outputs either. This command has become specific to outputting what our other packages will consume, while we have a separate set of scripts for syncing with Figma.

* Move these into the same folder

* Move the new token types and token-reading utils because they are more broadly useful

We will now need them in the CSS/Sass exporters

* Move some testing files and add some new tests

* In the flattened token files, keep the notation internally consistent

Let the Figma scripts do the conversion from `.` to `/`, but let our other exporters expect the alias notation to match the flattened keys

* Starting to write the new exporter for CSS vars

* Make it easier to test

by breaking it into functions that are easier to test and by only converting the files it finds rather than requiring all theme files to be present

* Now have the basic structure of the new exporter

Still need to resolve values more intelligently and generate variable names in a way that matches the old ones and conveys the correct hierarchy

* Structure of this script starting to take shape

* Add variable naming logic

* Tests for var naming

* New alias resolving logic and a bunch of comments

* Update other tests

* We can put both CSS and SCSS in the same file (and finish it)

* More snapshot tests, but now I see that layout is empty

The test data needs some layout tokens, but it gets complicated, because spacers are currently defined as plain numbers in Figma, whereas we'll probably want the JSON `$type` to be `dimension`. We need to intelligently convert the values between our JSON storage and Figma

* Add fiddly logic for converting from and to Figma's generic number type

Once we do our first down-sync from Figma, we could possibly start storing this translation information as meta-data inside the JSON tokens. It wouldn't get uploaded to Figma, but it could be manually maintained in our repository. Being able to store information that doesn't go to Figma, however, would require that we merge incoming data with our local JSON files instead of the current overwriting method.

Actually, if we're not storing that info in Figma, it's already in our tokens...why would I have to tell a dimension token that it's a dimension? If we don't need to be able to save the JSON based on only the information stored in Figma, then this isn't a problem at all. Maybe after the first down-sync we just need to update the down-sync operation to be a merge, and we look at the local (repository) token's `$type` and `$value` properties before converting from our Figma `NUMBER` variable. If `$type` is `dimension` then look at the unit of `$value` to determine how to translate from Figma.

* Turn that long commit message into comments so we can address later

* Get rid of extra quotes in CSS output and add support for percentages

This is so fiddly. I hate it.

* Update tokens package README

* Move test mock files into a less conspicuous and distracting place

* This list hierarchy got lost

* Fix minor TS errors

* Replace our old tokens with some test ones synced from Figma

We don't have the full set of tokens in Figma yet because we're in the process of trying to transfer the library file from one account to another, but we've got all the `System` variables, and I manually created a few variables in the `Theme` collection.

* Don't need these theme TS files anymore either

* Fix `"{spacer.4}px"` as a `$value` in the JSON

If a variable is an alias, we don't need to send it through our logic for converting pixel values to other units

* Add checks in the unit tests for that functionality and fix it again

My previous solution wasn't perfect because it didn't account for inheriting the token type from the aliased variable

* Round `ex` units

* More type refinement

* Create an example .env file

* Sync full set of variables from Figma

* Add duration conversion support. Pushing to Figma works!!

* WIP: Prompt and store decisions about ambiguous number types from Figma

* WIP: Add guessing back in. It works!

I now that I think about it again, I don't really think it's necessary to store the number type information in the `$extensions` when we can later infer from the `$type` and the converted `$value` string that it wrote, which should contain the unit. The only weird one would be `0` values that are also `dimensions`, but I say we always pretend those are pixel values.

* Use context clues from value and type instead of storing metadata

* Update unit tests and make sure we can get types for Figma from aliases

Now that we don't duplicate the type information on alias tokens, we have to resolve the aliases until we get to a real token with a type in order to send that information to Figma.

* Don't need these types anymore

* Add support for font weight tokens and update Figma

* Temporarily revert var printing and add some snapshot tests that will fail

They will fail because I've manually created the snapshots by copying the old CSS files so we can see the changes from this branch

* Whoops, I had used a different sort method to create the original snapshots

* Lowercase some variables in Figma

* Remove duplicate pagination variable in Figma

* Delete some more unused Figma variables

* Add some comments

* Use new official Figma API types package

Following the example of https://github.com/figma/variables-github-action-example/pull/19

* Some more doc comments

* Add a comment about the `ex` scale factor