[WNMGDS-2785] Fix first batch of Figma/CSS differences (#3107)
* Put the system tokens first in the css/scss output

In Sass, I'm pretty sure order of declaration matters, and we now reference system variables in theme variables. It also just feels more natural to have them up front if they're to be referenced later

* BREAKING: Use new button font-weight tokens instead of `--button__font-weight`

* Update refs to updated font variables and add other global inherited font variables

* Remove the need for `--inset__border-width` in healthcare theme

* Trying to get Choice to look right again

Filled in the missing gap token and fixed the medicare color

There's more to do on Choice, but I'm checking what I've got in for now

* Add a `choice/label/top-margin` variable to replace `choice__translateY`

* Fix choice border widths in medicare

* Remove `yarn prettier` package script because it conflicts with the base `yarn prettier` command

and causes an infinite loop when you try to run it. I've been using it to preview the token JSON changes before the pre-commit hook runs.

* Stop using the `--shadow-base` variable

* Add `label` vars to Figma and add support for `text-transform`

* Delete unused system variables from Figma

* Fix syncing `textTransform` to Figma

* Remove this from the diff

* BREAKING: Get rid of the `--hint__font-size` variable only used in medicare

This will mean a visual change for medicare where the hint text size now matches the error text size and defaults to 1rem

* Rename new `--choice-label__top-margin` variable to `--choice-label__margin-top`

* Rename the label offset variable

* Fix bad typography variable references and disuse of font weight vars

* Fix incorrect token values for medicare heading sizes

* Fix 2xl headings not getting their font weight from tokens

* BREAKING: Add missing theme-level "normal" and "bold" font weight

Removing `--font-weight-semibold` and `--font-weight-light` in the process. We don't use them in our other tokens. The semibold one is used in a utility class only. I've hard-coded the value into the utility class until it's deemed time to deprecate that utility class.

* User error inputting the updated medicare 5xl value in Figma

* More fixes to medicare font sizes. Now all the size-related VRTs pass

The only remaining VRT failures for typography are Link changes

* Forgot to rename this in the CSS

* BREAKING: Deprecate semibold utility class and remove semibold Open Sans from includes

* Update VRTs for the medicare hint text, whose size we changed

* Add deprecation styles to removed `ds-u-font-family--semibold`