WIP: Well, I thought this was a good idea
Turns out that medicare does use a different value for the max-width, and that's `unset`. Having pixel values and `unset` doesn't work in Figma, so I think a better option there is to make it a boolean. However, that boolean still has to play nice with the variable we told people in the docs to use to set the max width (`--usa-banner__max-width`).