[WNMGDS-2785] Update USA Banner with Figma changes (#3112)
* Use new gutter variable

I'm already thinking I'm going to reverse this and just delete it from Figma. You'll see

* Revert "Use new gutter variable"

This reverts commit ea0c2a95fa0d81a383523bb93a784abc5c2e77b4.

* WIP: Well, I thought this was a good idea

Turns out that medicare does use a different value for the max-width, and that's `unset`. Having pixel values and `unset` doesn't work in Figma, so I think a better option there is to make it a boolean. However, that boolean still has to play nice with the variable we told people in the docs to use to set the max width (`--usa-banner__max-width`).

* Use a boolean value for the theme data but allow apps to override the max width

* Fix medicare link color and get rid of two more variables

Updated the VRT ref images too