[NO-TICKET] Fix broken doc-site build (#3120)
Don't print any of the boolean values on the doc site

Those boolean tokens are mostly internal and are not robust enough to set to different values. They shouldn't be shown in the docs as configuration options. This fixes build because I made the CSS value function to throw an error if we try to get the value of a boolean token.