Using `useState` instead of `useReducer` doesn't work either
Bugger, my call to a `userReducer` dispatch function doesn't work in my `useEffect` cleanup because of a React bug/feature regarding calling state-setting functions within cleanup functions. I need to reference state in the cleanup function, though, so I can undo what we did to the body element. This is in regards to `useBodyScrollPrevention`.

Even if I implement it with `useState` instead of `useReducer`, if the `useEffect` call doesn't have state in its dependency array, the version of the cleanup function that gets executed is the one that was created on the initial render when the state was first initialized. That means the current opened state and its info won't be available on in that cleanup function. How do I know when it's being unmounted prematurely but still have access to the current state?