[WNMGDS-2811] Fix body scroll freeze persisting after dialog unmounts (#3140)
* WIP: Debugging

For some reason when we're unmounting the component, it thinks it's in a closed state even though the close function was never called. Steps to reproduce:

1. `yarn build:core && yarn build:docs && yarn serve:docs`
2. Open http://localhost:9000/
3. Navigate to the Alert component page
4. Scroll down to the maturity section and click the new button to open a dialog
5. Notice in the console that it says "opening"
6. Hit the browser back button
7. Notice that the console never says "closing", but the local state of `useBodyScrollPrevention` seems to be closed

* Reproduce it in Storybook, where it's easier to debug

* Reproduce it in Storybook for real

Just having it in another tab didn't actually de-render it because it was being rendered in the root story component function even if it wasn't visible.

* Using `useState` instead of `useReducer` doesn't work either

Bugger, my call to a `userReducer` dispatch function doesn't work in my `useEffect` cleanup because of a React bug/feature regarding calling state-setting functions within cleanup functions. I need to reference state in the cleanup function, though, so I can undo what we did to the body element. This is in regards to `useBodyScrollPrevention`.

Even if I implement it with `useState` instead of `useReducer`, if the `useEffect` call doesn't have state in its dependency array, the version of the cleanup function that gets executed is the one that was created on the initial render when the state was first initialized. That means the current opened state and its info won't be available on in that cleanup function. How do I know when it's being unmounted prematurely but still have access to the current state?

* Well, no FSMs, no bells or whistles. It works

* Add more tests and clean up

* Revert reproduction scenario in Storybook and the doc site