[WNMGDS-2808] Fix Firefox scrolling pages to the top when a dialog opens (#3142)
* Fix Firefox scrolling pages to the top when a dialog opens in race condition

On one of the apps, opening a dialog in Firefox would scroll the page to the top. We weren't able to reproduce this in our Storybook, but we could see the behavior that caused it by disabling our code that applied the body classes and properties that freeze it. The polyfill styles by themselves were causing the page to scroll to the top in Firefox. In the race condition, the page would scroll to the top before our JavaScript could poll the document to find its scroll position, so the scroll position was always zero.

We don't need these polyfill styles anymore. We have to keep the skeleton of a JavaScript polyfill around to support testing frameworks, but those don't need any CSS. The dialog element has Baseline 2022 support.

* Oh, lol, actually get rid of the styles