This script isn't needed anymore and could be confusing.
Our current method of testing in a downstream application is to take a project that already has it installed and replace its `node_modules/@cmsgov/<PACKAGE_NAME>/dist` folder with the dist folder that we build in our project. That works in most cases, but sometimes you'll also need dependencies. In that case, this script isn't going to work anyway, and we might need to publish a beta version if something like `npm link` doesn't work.