Undo the change to make descriptions with links React elements
This was fine on our docs page with our custom attributes table, but Storybook crashed trying to render these values in story controls. The error was completely unhelpful and the stack trace was all through minified code, so I had to guess and test to find the problem here. Now we know these can only be strings.