[WNMGDS-2843] Add `ds-text-field` component (#3172)
* Add basics of new `ds-text-field` component

* Add autofocus since we added it to ds-droppdown

* Add some unit tests

* Make the built-in option strings easier on the eyes and add a label-mask test

* Add it to the index

* Make sure it renders as block

* Add browser test snapshots

* Fix attribute docs according to feedback

* Fix possibility of passing both `autofocus` and `autoFocus` to Preact component

* Added description to requirement-label

* Rewrite label mask prop/attr descriptions and fix broken links in Storybook

* Undo the change to make descriptions with links React elements

This was fine on our docs page with our custom attributes table, but Storybook crashed trying to render these values in story controls. The error was completely unhelpful and the stack trace was all through minified code, so I had to guess and test to find the problem here. Now we know these can only be strings.

* Add support for multiline fields and a unit test for it

* Don't use the name "wrapper" because it can be confused with the WC wrapper

And honestly we might not even need this element anymore. Sarah brought that up, and it's worth considering.

* Accidentally left this in, which messes up VRTs

* Add snapshots for the new multi-line story