[NO-TICKET] Use the external link component in doc site (#2522)
* Use the external link component as-is for all external links on doc site

Note that we're going to need to refine the component's styles in order for it to work well

* Use link after talking to a11y experts, and update styles

Our argument for using a link is that even though it doesn't perform the navigation directly (normal a11y rule for using a button instead of a link), the intent of the element is to perform navigation, which makes it a link. This also allows desktop users to hover over the link and see the address, and it allows some other standard browser behaviors. It is also way easier to style to look like a link, because it is a link. We realized we were going to have to fix the styles anyway because its primary use case is to be within blocks of text.

* Update unit tests and snapshots

* Be more discerning about URLs that we slap the external link pattern on