Add a comment at the top of the template file
This should give people a clue as to what this file is about, and I think the content of this comment is such that it wouldn't look out of place copied into different projects. It even gives a clue to people who might be tempted to modify those example-project copies directly to find out where it's being copied _from_ first.