[WNMGDS-2999] Add a shared web components example template and new example project (#3267)
* Move the common web components template into its own file

* Added a new script to replace template content in examples

It works, but it doesn't indent. I'm not sure if I'm going to run prettier on it or just try to indent manually to the place where the starting comment is indented.

* Indent!

We don't have prettier configured to format astro files right now, and I want it to be flexible enough for all file types even if we do configure it for astro

* Add some code comments

* Create a package script entry (and rename it)

* Run the script and save the results

* Add an example project that imports web components from the different themes

* Apply the template to the new project and fix the project

* Create an example build command wrapper that observes `SMOKE` env var

and doesn't build the `astro-themes` project if it's present.

* Add new tests and update snapshots

* Update examples readme to describe the new example project

* Fix duplicate package names!

* Fix JSX-specific syntax in template that caused silent failures in cdn examples

* Update VRT reference images that showed malformed choice-list-item

This was due to a JSX-style self-closing tag. Note that the Autocomplete does not render because this is using the latest CDN version, which does not have the new Autocomplete web component!

* Update the examples by running `apply-example-templates` when we build examples

This will hopefully keep the template and examples from getting out of sync.

Note that when I ran this script the examples got updated. One of my previous commits must have applied Prettier formatting to the template file, and that resulted in changes now.

* Add a comment at the top of the template file

This should give people a clue as to what this file is about, and I think the content of this comment is such that it wouldn't look out of place copied into different projects. It even gives a clue to people who might be tempted to modify those example-project copies directly to find out where it's being copied _from_ first.